changed datatype int8_t -> int to avoid cast problems

+++ include/options.hpp
@@ -349,11 +349,11 @@ public:
 
 	uint32_t minoccur = 0; // TODO: add to cmd options. Min number of k-mer occurrences in the DB to use for matching. See 'index.lookup_tbl[kmer_idx].count'
 
-	int8_t match = 2; // '--match' SW score (positive integer) for a match
-	int8_t mismatch = -3; // '--mismatch' SW penalty (negative integer) for a mismatch
+	int match = 2; // '--match' SW score (positive integer) for a match
+	int mismatch = -3; // '--mismatch' SW penalty (negative integer) for a mismatch
 	long gap_open = 5; // '--gap_open' SW penalty (positive integer) for introducing a gap
 	long gap_extension = 2; // '--gap_ext' SW penalty (positive integer) for extending a gap
-	int8_t score_N = 0; // '-N' SW penalty for ambiguous letters (N's)
+	int score_N = 0; // '-N' SW penalty for ambiguous letters (N's)
 
 	double evalue = -1.0; // '-e' E-value threshold
 	double align_id = -1.0; // OTU-picking option: minimum %%id to keep alignment
